backend/requirements.txt
django==2.2.17
djangorestframework==3.12.2
Celery==5.0.2
# django-phone-field==1.8.1
# django-countries==6.1.3
phonenumbers==8.12.13
django-address==0.2.5
django-money==1.1
psycopg2-binary
django-phonenumber-field==5.0.0
django-cors-headers==3.5.0
django-filter==2.4.0
